Sprout gift boxes!

+++ src/main/java/mysticmods/roots/api/datamap/DataMaps.java
@@ -66,8 +66,8 @@ public class DataMaps {
   public static final DataMapType<PlayerCondition, CanonicalRepresentation> PLAYER_CONDITION_CANONS = DataMapType.builder(RootsAPI.rl("player_condition_canons"), RootsRegistries.Keys.PLAYER_CONDITIONS, CanonicalRepresentation.CODEC)
       .synced(CanonicalRepresentation.CODEC, false)
       .build();
-  public static final DataMapType<Item, Float> SPROUT_BREEDING_ITEM_CHANCE = DataMapType.builder(RootsAPI.rl("sprout_breeding_item_chance"), Registries.ITEM, Codec.FLOAT)
-      .synced(Codec.FLOAT, false)
+  public static final DataMapType<Item, Integer> SPROUT_BREEDING_ITEM_CHANCE = DataMapType.builder(RootsAPI.rl("sprout_breeding_item_chance"), Registries.ITEM, Codec.INT)
+      .synced(Codec.INT, false)
       .build();
 
   @SubscribeEvent
